requirements.txt
click>=8.0.0
pyyaml>=6.0.0
jsonschema>=4.0.0
jinja2>=3.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
